test(cli): split config function tests
